Start replacing printpdf with pdf-writer

+++ src/elements/align_preferred_height_bottom.rs
@@ -90,7 +90,7 @@ impl<'a, E: Element> Element for AlignPreferredHeightBottom<'a, E> {
 #[derive(Debug)]
 struct Layout {
     breaks: u32,
-    y_offset: f64,
+    y_offset: f32,
     size: ElementSize,
 }
 
@@ -98,10 +98,10 @@ impl<'a, E: Element> AlignPreferredHeightBottom<'a, E> {
     fn layout(
         &self,
         width: WidthConstraint,
-        first_height: f64,
-        full_height: Option<f64>,
+        first_height: f32,
+        full_height: Option<f32>,
         preferred_breaks: u32,
-        preferred_height: f64,
+        preferred_height: f32,
     ) -> Layout {
         let height_available = full_height.unwrap_or(first_height);
